[base][drape] Fix `deleted operator <<` of char32_t on C++20
Commit f05be9bf713a6ad7e686695de89296b728a26176,
changed `UniChar` from `char32_t` to `uint32_t`.
Unlike `uint32_t` the `char32_t` has explicitly
deleted `<<` operator in C++20, and this operator
has been in use on `UniChar`.

+++ drape/drape_tests/bidi_tests.cpp
@@ -32,11 +32,11 @@ UNIT_TEST(Bidi_Print)
   {
     auto const s = strings::MakeUniString(utf8);
     for (auto const ucp : s)
-      std::cout << std::hex << ucp << ", ";
+      std::cout << std::hex << static_cast<uint32_t>(ucp) << ", ";
 
     std::cout << std::endl;
     for (auto const ucp : bidi::log2vis(s))
-      std::cout << std::hex << ucp << ", ";
+      std::cout << std::hex << static_cast<uint32_t>(ucp) << ", ";
 
     std::cout << std::endl;
   };
